jessie 2014.7.5 continue reading events even after getting one with wrong tag
the event polling loop should return None if no events are available,
the event if we find one we 'subscribed' to (has a matching tag),
or keep reading if we got an event intended for another client.

Previously it returned None on reading an event with the wrong tag;
this caused the salt cli script to stop listening for minion
returns early.

This bug has been fixed in 2015.8.x due to a refactoring of the code.

Change-Id: I1b860f30f30eb8b9e352704132873f1942976a62