jessie 2014.7.5 continue reading events even after getting one with wrong tag
the event polling loop should return None if no events are available,
the event if we find one we 'subscribed' to (has a matching tag),
or keep reading if we got an event intended for another client.

Previously it returned None on reading an event with the wrong tag;
this caused the salt cli script to stop listening for minion
returns early.

This bug has been fixed in 2015.8.x due to a refactoring of the code.

Revert "Revert "jessie 2014.7.5 continue reading events even after getting one with wrong tag""
This reverts commit 52b7c3f1fac1c6b34420f8b033938726b3596ea6.
resubmitting to go through gerrit

Change-Id: I0a0f23166b5d6669eab0a945247723b2858b6e45